saf: Rewrite SAF with view types
Provide better typing and cleaner signatures throughout by keeping
the shape information on the SAFs where it belongs, at the cost of
some boilerplate to provide a new iterator type and a bit of code
duplication over owned and view types. This can be gradually
improved as GAT/TAIT/DST stabilises.