cli/run: Use builder
The builder is now in charge of defaults and validation. Note also
that the block to block size conversion subtly changes in this
connection: we do not attempt to calculate a block size that
exactly fits the desired number of blocks, since this will often
fail anyway. Instead, we simply let `blocks = sites / block_size`,
and accept that this can be off by a little bit relative to the
desired number of blocks.