Rewrite block (parallel) iterators
Previously, these iterators required a block size to iterate over,
and the last block would be smaller if the total number of sites was
not divided evenly by the block size. This made it cumbersome and
imprecise to specify a number of blocks instead of a block size.

Block iterators now take either a block size or a number of blocks.
Given a block size, the behaviour is unchanged from before.
Given a number of blocks, this number will now be exact,
and all blocks will be within one site of each other in size.

For practical usage of winsfs, this should make no difference, but
it makes its behaviour less surprising and enables finer control
if needed.